streamline keyvaluereducer and arrayreducer
