+++ src/function/game-folder/state.ts
@@ -1,4 +1,3 @@
-
 import { atom, useAtomValue } from 'jotai';
 import { loadable } from 'jotai/utils';
 import { exists } from '@tauri-apps/api/fs';
