+++ src/function/global/ArrayReducer.ts
@@ -1,0 +1,8 @@
+export function ArrayReducer<Type>() {
+  return (
+    _state: ArrayReducerState<Type>,
+    newState: ArrayReducerArgs<Type>,
+  ) => [...newState];
+}
+export type ArrayReducerArgs<Type> = Type[];
+export type ArrayReducerState<Type> = Type[];
